database/postgres_sync.py
import os
import psycopg2
from psycopg2 import sql

class PostgresSync:
    """
    PostgreSQL Sync: Handles hard metrics like price history, volume, and sector stats.
    """
    def __init__(self, connection_url: str = None):
        self.connection_url = connection_url or os.getenv("POSTGRES_URL")
        self.conn = None
        if self.connection_url:
            self.conn = psycopg2.connect(self.connection_url)

    def initialize_schema(self):
        if not self.conn:
            print("Postgres connection not available.")
            return

        with self.conn.cursor() as cur:
            # Create a sample table for market metrics
            cur.execute("""
                CREATE TABLE IF NOT EXISTS market_metrics (
                    id SERIAL PRIMARY KEY,
                    symbol VARCHAR(20) NOT NULL,
                    price DECIMAL(18, 2),
                    volume BIGINT,
                    timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
                );
            """)
            self.conn.commit()

    def insert_metric(self, symbol: str, price: float, volume: int):
        if not self.conn: return
        with self.conn.cursor() as cur:
            cur.execute(
                "INSERT INTO market_metrics (symbol, price, volume) VALUES (%s, %s, %s)",
                (symbol, price, volume)
            )
            self.conn.commit()

if __name__ == "__main__":
    # sync = PostgresSync()
    # sync.initialize_schema()
    pass
